Move `neptune.common` - PART 2 (#1693)
