+++ pyproject.toml
@@ -208,36 +208,35 @@ module = [
     "neptune.attributes.series.string_series",
     "neptune.attributes.sets.string_set",
     "neptune.attributes.utils",
-    "neptune.common.backends.utils",
-    "neptune.common.exceptions",
+    "neptune.internal.exceptions",
     "neptune.internal.utils.git_info",
-    "neptune.common.hardware.cgroup.cgroup_filesystem_reader",
-    "neptune.common.hardware.cgroup.cgroup_monitor",
-    "neptune.common.hardware.gauges.cpu",
-    "neptune.common.hardware.gauges.gauge",
-    "neptune.common.hardware.gauges.gauge_factory",
-    "neptune.common.hardware.gauges.gpu",
-    "neptune.common.hardware.gauges.memory",
-    "neptune.common.hardware.gpu.gpu_monitor",
-    "neptune.common.hardware.metrics.metric",
-    "neptune.common.hardware.metrics.metrics_container",
-    "neptune.common.hardware.metrics.metrics_factory",
-    "neptune.common.hardware.metrics.reports.metric_reporter",
-    "neptune.common.hardware.metrics.reports.metric_reporter_factory",
-    "neptune.common.hardware.metrics.service.metric_service",
-    "neptune.common.hardware.metrics.service.metric_service_factory",
-    "neptune.common.hardware.resources.gpu_card_indices_provider",
-    "neptune.common.hardware.resources.system_resource_info",
-    "neptune.common.hardware.resources.system_resource_info_factory",
-    "neptune.common.hardware.system.system_monitor",
+    "neptune.internal.hardware.cgroup.cgroup_filesystem_reader",
+    "neptune.internal.hardware.cgroup.cgroup_monitor",
+    "neptune.internal.hardware.gauges.cpu",
+    "neptune.internal.hardware.gauges.gauge",
+    "neptune.internal.hardware.gauges.gauge_factory",
+    "neptune.internal.hardware.gauges.gpu",
+    "neptune.internal.hardware.gauges.memory",
+    "neptune.internal.hardware.gpu.gpu_monitor",
+    "neptune.internal.hardware.metrics.metric",
+    "neptune.internal.hardware.metrics.metrics_container",
+    "neptune.internal.hardware.metrics.metrics_factory",
+    "neptune.internal.hardware.metrics.reports.metric_reporter",
+    "neptune.internal.hardware.metrics.reports.metric_reporter_factory",
+    "neptune.internal.hardware.metrics.service.metric_service",
+    "neptune.internal.hardware.metrics.service.metric_service_factory",
+    "neptune.internal.hardware.resources.gpu_card_indices_provider",
+    "neptune.internal.hardware.resources.system_resource_info",
+    "neptune.internal.hardware.resources.system_resource_info_factory",
+    "neptune.internal.hardware.system.system_monitor",
     "neptune.internal.oauth",
-    "neptune.common.patches.bravado",
-    "neptune.common.storage.datastream",
-    "neptune.common.storage.storage_utils",
+    "neptune.internal.patches.bravado",
+    "neptune.internal.storage.datastream",
+    "neptune.internal.storage.storage_utils",
     "neptune.internal.utils.utils",
     "neptune.internal.warnings",
-    "neptune.common.websockets.reconnecting_websocket",
-    "neptune.common.websockets.websocket_client_adapter",
+    "neptune.internal.websockets.reconnecting_websocket",
+    "neptune.internal.websockets.websocket_client_adapter",
     "neptune.exceptions",
     "neptune.handler",
     "neptune.integrations.python_logger",
@@ -312,7 +311,7 @@ module = [
     "neptune.api.requests_utils",
     "neptune.version",
     "neptune.internal.container_type",
-    "neptune.common.patches",
+    "neptune.internal.patches",
     "neptune.internal.utils",
 ]
 ignore_errors = "True"
