+++ src/neptune/internal/backends/hosted_file_operations.py
@@ -48,19 +48,6 @@ from requests import (
     Response,
 )
 
-from neptune.common.hardware.constants import BYTES_IN_ONE_MB
-from neptune.common.storage.datastream import (
-    FileChunk,
-    FileChunker,
-    compress_to_tar_gz_in_memory,
-)
-from neptune.common.storage.storage_utils import (
-    AttributeUploadConfiguration,
-    UploadEntry,
-    normalize_file_name,
-    scan_unique_upload_entries,
-    split_upload_files,
-)
 from neptune.exceptions import (
     FileUploadError,
     MetadataInconsistency,
@@ -81,6 +68,17 @@ from neptune.internal.exceptions import (
     InternalClientError,
     NeptuneException,
     UploadedFileChanged,
+)
+from neptune.internal.hardware.constants import BYTES_IN_ONE_MB
+from neptune.internal.storage import (
+    AttributeUploadConfiguration,
+    FileChunk,
+    FileChunker,
+    UploadEntry,
+    compress_to_tar_gz_in_memory,
+    normalize_file_name,
+    scan_unique_upload_entries,
+    split_upload_files,
 )
 from neptune.internal.utils import (
     get_absolute_paths,
