+++ src/neptune/internal/hardware/metrics/reports/metric_reporter_factory.py
@@ -13,3 +13,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+from neptune.internal.hardware.metrics.reports.metric_reporter import MetricReporter
+
+
+class MetricReporterFactory(object):
+    def __init__(self, reference_timestamp):
+        self.__reference_timestamp = reference_timestamp
+
+    def create(self, metrics):
+        return MetricReporter(metrics=metrics, reference_timestamp=self.__reference_timestamp)
