+++ src/neptune/internal/patches/__init__.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2019, Neptune Labs Sp. z o.o.
+# Copyright (c) 2022, Neptune Labs Sp. z o.o.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,3 +13,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+__all__ = ["apply_patches"]
+
+from neptune.internal.patches.bravado import patch as bravado_patch
+
+patches = [bravado_patch]
+
+
+# Apply patches when importing a patching module
+# Should be called before usages of patched objects
+def apply_patches():
+    for patch in patches:
+        patch()
