+++ src/neptune/internal/utils/source_code.py
@@ -23,7 +23,7 @@ from typing import (
 )
 
 from neptune.attributes import constants as attr_consts
-from neptune.common.storage.storage_utils import normalize_file_name
+from neptune.internal.storage import normalize_file_name
 from neptune.internal.utils import (
     does_paths_share_common_drive,
     get_absolute_paths,
