+++ src/neptune/internal/websockets/websockets_factory.py
@@ -20,7 +20,7 @@ from typing import Optional
 
 from requests_oauthlib import OAuth2Session
 
-from neptune.common.websockets.reconnecting_websocket import ReconnectingWebsocket
+from neptune.internal.websockets.reconnecting_websocket import ReconnectingWebsocket
 
 
 class WebsocketsFactory:
